Port to JymKit (#4)
Porting this environment to a [JymKit](https://github.com/ponseko/jymkit) environment, simplifying the codebase and focussing it only on the environment. 

At the time of release, this updated version may perform a few minutes slower due to the way how JymKit deals with a mix of homogeneous and heterogeneous agents. This should be adressed in later releases of JymKit.